fix(Kandel): PricePoints cannot be more than 32bits since offerId is only 32bits (#109)

+++ script/strategies/kandel/KandelShutdown.s.sol
@@ -22,7 +22,7 @@ contract KandelShutdown is Deployer {
     IERC20 quote = kdl.QUOTE();
     uint baseDecimals = base.decimals();
     uint quoteDecimals = quote.decimals();
-    (,,, uint112 length) = kdl.params();
+    (,,, uint32 length) = kdl.params();
 
     uint baseBalance = base.balanceOf(broadcaster());
     uint quoteBalance = quote.balanceOf(broadcaster());
